Switch to impl & wrapper for gvim
So that we can target a specific python virtual env to use when running
the script.

Switch call sites to use gvim-wrapper.

The inclusion of wmctrl is to focus the gvim window after opening a file
in it, as it doesn't always bring focus on xubuntu 13.10, I think.

I have no idea why I've added +23 to the line_number.